+++ fb.h
@@ -3,6 +3,6 @@
 #ifndef __FB_H__
 #define __FB_H__
 
-extern uint32_t *FB_Init(int *width, int *height, int *bpp);
+extern uint8_t *FB_Init(int *width, int *height, int *bpp);
 
 #endif  /* __FB_H__ */
